Fix testing environment settings.
Previously, a `.env.test` file was present on the root of the project, and the test Docker container had a `APP` environment variable that was supposed to force Symfony to use this file instead of the default `.env` one. It seems that this technique was not working quite well… A clearer one was thus implemented that work in the same way as for dev container: we link the `.env` file to use at the correct place with a Docker volume.